FACES-2977 Develop tests for the Portlet and Bridge:InputFile category of the JSF Showcase (Portlet) (code review)
